Move install command implementation to install.go
