+++ plugin/install_test.go
